refactor: create a module for computer vision
